+++ NEWS.md
@@ -1,7 +1,11 @@
+# citr 0.3.0
 
 - New function `tidy_bib_file()` removes duplicate and unneeded entries from a Bib(La)Tex-file.
-- Added `encoding` option to functions that read Bib(La)TeX-files and a corresponding global option. Option is also used to read parent documents in `insert_citation()` (reported by @sammo3182 #11)
+- Added `encoding` option to functions that read Bib(La)TeX-files and a corresponding global option. (reported by @sammo3182 #11)
 - Error message in `insert_citation()` are printed to the dialog instead of the console (reported by @sammo3182 #11)
+- RStudio addin
+    - Access to Zotero group libraries (#9)
+    - New settings tab
 
 # citr 0.2.0